refactor: 작성한 Network 모듈로 네트워크 기능 대체
