+++ iOS/FlipMate/FlipMate/Application/DIContainer/AppDIContainer.swift
@@ -6,10 +6,11 @@
 //
 
 import UIKit
+import Core
+import Network
 
 final class AppDIContainer {
-    lazy var keychainManager: KeychainManageable = KeychainManager()
-    lazy var provider: Provider = Provider(urlSession: URLSession.shared, keychainManager: keychainManager)
+    lazy var provider: Provider = Provider(urlSession: URLSession.shared, keychainManager: KeychainManager())
     lazy var categoryManager: CategoryManageable = CategoryManager(categories: [])
     lazy var userInfoManager: UserInfoManageable = UserInfoManager()
     
@@ -27,8 +28,7 @@ final class AppDIContainer {
         let dependencies = LoginDIContainer.Dependencies(
             provider: provider,
             categoryManager: categoryManager,
-            userInfoManager: userInfoManager,
-            keychainManager: keychainManager
+            userInfoManager: userInfoManager
         )
         
         return LoginDIContainer(dependencies: dependencies)
