+++ iOS/FlipMate/FlipMate/Application/DIContainer/CategoryDIContainer.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+import Network
 
 final class CategoryDIContainer: CategoryFlowCoordinatorDependencies {
     struct Dependencies {
