+++ iOS/FlipMate/FlipMate/Application/DIContainer/ChartDIContainer.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+import Network
 
 final class ChartDIContainer: ChartFlowCoordinatorDependencies {
     struct Dependencies {
