+++ iOS/FlipMate/FlipMate/Application/DIContainer/TabBarDIContainer.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+import Network
 
 final class TabBarDIContainer: TabBarFlowCoordinatorDependencies {
