+++ iOS/FlipMate/FlipMate/Application/DIContainer/TimerSceneDIContainer.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+import Network
 
 final class TimerSceneDIContainer: TimerFlowCoordinatorDependencies {
     struct Dependencies {
