+++ iOS/FlipMate/FlipMate/Data/Network/CategoryEndpoints.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import Network
 
 struct CategoryEndpoints {
