--- iOS/FlipMate/FlipMate/NetworkService/HTTPHeader.swift
@@ -1,13 +1,0 @@
-//
-//  HTTPHeader.swift
-//  FlipMate
-//
-//  Created by 임현규 on 2023/11/22.
-//
-
-import Foundation
-
-struct HTTPHeader {
-    var value: String
-    var field: String
-}
