--- iOS/FlipMate/FlipMate/NetworkService/HTTPMethod.swift
@@ -1,15 +1,0 @@
-//
-//  HttpMethod.swift
-//  FlipMate
-//
-//  Created by 신민규 on 11/21/23.
-//
-
-import Foundation
-
-enum HTTPMethod: String {
-    case get = "GET"
-    case post = "POST"
-    case delete = "DELETE"
-    case patch = "PATCH"
-}
