--- iOS/FlipMate/FlipMate/NetworkService/Responsable.swift
@@ -1,12 +1,0 @@
-//
-//  Responsable.swift
-//  FlipMate
-//
-//  Created by 신민규 on 11/21/23.
-//
-
-import Foundation
-
-protocol Responsable {
-    associatedtype Response: Decodable
-}
